Move items to modules in which they best belong
